+++ src/main/agent/optimizer.ts
@@ -1,8 +1,16 @@
-import { type AgentProfile } from '@common/types';
+import { type AgentProfile, ToolApprovalState } from '@common/types';
 import { CacheControl } from 'src/main/agent/llm-provider';
 import { cloneDeep } from 'lodash';
 import { type CoreUserMessage, type CoreMessage, type ToolContent, type ToolResultPart } from 'ai';
-import { AIDER_TOOL_GROUP_NAME, AIDER_TOOL_RUN_PROMPT, TODO_TOOL_GET_ITEMS, TODO_TOOL_GROUP_NAME, TOOL_GROUP_NAME_SEPARATOR } from '@common/tools';
+import {
+  AIDER_TOOL_GROUP_NAME,
+  AIDER_TOOL_RUN_PROMPT,
+  POWER_TOOL_AGENT,
+  POWER_TOOL_GROUP_NAME,
+  TODO_TOOL_GET_ITEMS,
+  TODO_TOOL_GROUP_NAME,
+  TOOL_GROUP_NAME_SEPARATOR,
+} from '@common/tools';
 
 import logger from '../logger';
 
@@ -56,6 +64,12 @@ const addImportantReminders = (profile: AgentProfile, userRequestMessageIndex: n
 
   if (!profile.autoApprove) {
     dontForgets.push('Before making any changes, present the plan and wait for my approval.');
+  }
+
+  if (profile.usePowerTools && profile.toolApprovals[`${POWER_TOOL_GROUP_NAME}${TOOL_GROUP_NAME_SEPARATOR}${POWER_TOOL_AGENT}`] !== ToolApprovalState.Never) {
+    dontForgets.push(
+      `Prefer using \`${POWER_TOOL_GROUP_NAME}${TOOL_GROUP_NAME_SEPARATOR}${POWER_TOOL_AGENT}\` tool to delegate analysis and complex tasks to a sub-agent.`,
+    );
   }
 
   if (dontForgets.length === 0) {